FAC: Move execution to a new module
